backend/requirements.txt
Flask
flask-cors
PyMuPDF
Pillow
gunicorn
pymongo
pytesseract
watchdog
apscheduler
requests
python-dotenv
